oci: Add support for CRI-O
Relying on latest changes from virtcontainers, this patch modifies
oci.go so that it can be used by every command implemented in the
runtime.

Notice that it factorizes an existing function so that we can prevent
the code to call several times into ListPod() which cost us some
performances. Indeed, looping over the list of pods and containers
is relatively high cost because we could potentially loop over
thousands of containers/pods and we access the local storage for this.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>